fix(postgres): bugs de migración encontrados conectando a Neon
Arregla los 500 que daban los GET tras conectar el API a PostgreSQL en Neon.
Login funcionaba pero todos los demás endpoints fallaban con errores de
GETDATE inexistente, ISNULL mal traducido, tipos boolean vs int en COALESCE,
literales "1" para Anulada en lugar de TRUE, y parámetros NULL sin tipo que
PG no podía inferir ("42P08: could not determine data type").

Cambios principales:
- AuthMiddleware: GETDATE() -> NOW()
- RequirePermiso: ISNULL(...) -> COALESCE(...)
- InventarioController: COALESCE(TieneImagen, 0) -> COALESCE(TieneImagen, FALSE),
  casts ::text/::int en filtros opcionales del paginado.
- NotasCreditoController: anular Boleta usaba "1" -> TRUE.
- AjustesController, CambioPreciosController, ReportesController:
  casts ::date/::timestamp/::int/::text en todos los filtros @Desde/@Hasta/
  @CategoriaId/@ProveedorId/@Bodega/@Categoria opcionales.

Verificación: 25/25 endpoints (CRUD + 12 reportes) responden 200 OK
contra Neon PostgreSQL con login válido.

+++ Controllers/CambioPreciosController.cs
@@ -36,7 +36,7 @@ namespace ClaumanAPI.Controllers
                        COALESCE(i.PrecioMayor, 0)  AS PrecioMayor
                 FROM Inventario i
                 LEFT JOIN Categorias c ON c.Id = i.CategoriaId
-                WHERE (@CategoriaId IS NULL OR i.CategoriaId = @CategoriaId)
+                WHERE (@CategoriaId::int IS NULL OR i.CategoriaId = @CategoriaId::int)
                 ORDER BY i.Descripcion";
 
             var cmd = new NpgsqlCommand(sql, conexion);
@@ -94,7 +94,7 @@ namespace ClaumanAPI.Controllers
                 var sql = $@"
                     UPDATE Inventario
                     SET {set}
-                    WHERE (@CategoriaId IS NULL OR CategoriaId = @CategoriaId)";
+                    WHERE (@CategoriaId::int IS NULL OR CategoriaId = @CategoriaId::int)";
 
                 var cmd = new NpgsqlCommand(sql, conexion, tx);
                 cmd.Parameters.AddWithValue("@Factor",      factor);
